+++ src/components/Layout/Footer.js
@@ -1,19 +1,24 @@
 import React from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faEnvelope } from "@fortawesome/free-solid-svg-icons";
 import { faLinkedin, faGithub } from "@fortawesome/free-brands-svg-icons";
 
 const Footer = () => (
-  <footer className="text-center justify-center font-abc py-9 bg-cream">
-    <p>Developed by Ken Franco</p>
+  <footer className="text-center justify-center font-abc py-9 bg-red3 text-cream text-xl">
+    <p className="mb-5">© 2023 Ken Franco</p>
     <a
-      className="hover:text-red1"
+      className="hover:text-gold"
       href="https://www.linkedin.com/in/kenneth-laurenz-franco-162b94200/"
     >
       <FontAwesomeIcon className="mr-5" icon={faLinkedin} />
     </a>
 
-    <a className="hover:text-red1" href="https://github.com/KennethFranco">
-      <FontAwesomeIcon icon={faGithub} />
+    <a className="hover:text-gold" href="https://github.com/KennethFranco">
+      <FontAwesomeIcon className="mr-5" icon={faGithub} />
+    </a>
+
+    <a className="hover:text-gold" href="mailto:kennethlaurenzfranco@gmail.com">
+      <FontAwesomeIcon icon={faEnvelope} />
     </a>
   </footer>
 );