+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog Format
 - [Initials of Dev] [#issue number](issue link on GitHub) {issue title}
 -->
 
+- [KF] [#17](https://github.com/KennethFranco/portfolio-v2/issues/17) home page(feature): add projects carousel
 - [KF] [#15](https://github.com/KennethFranco/portfolio-v2/issues/15) app(feature): apply 230317 ui changes
 - [KF] [#13](https://github.com/KennethFranco/portfolio-v2/issues/13) app(feature): clean up files
 - [KF] [#11](https://github.com/KennethFranco/portfolio-v2/issues/11) home page(feature): add skills frame